Settings_Extrapolation: moved function to it's own module so individual proxifiers can call.
